+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added support for query `PeerTotalSent`
 - Added support for query `PeerList`
 - Added support for query `PeerStats`
+- Added support for query `PeerVersion`
 
 ## 3.0.0
 - Stronger typing for various places in the API.